device-posture/device-posture-clear.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';

promise_test(async (t) => {
  const originalPosture = navigator.devicePosture.type;
  const posture = originalPosture ? 'folded' : 'continuous';

  await test_driver.set_device_posture(posture);
  assert_equals(navigator.devicePosture.type, posture);

  await test_driver.clear_device_posture();
  assert_equals(navigator.devicePosture.type, originalPosture);
}, 'Tests that device posture override can be removed.');
</script>
